Refactor services, add TODO list

+++ src/main/java/ua/alexkras/hotel/repository/ApartmentRepository.java
@@ -16,12 +16,12 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ApartmentRepository extends PagingAndSortingRepository<Apartment,Long> {
-    Optional<Apartment> findApartmentById(Integer id);
+    Optional<Apartment> findApartmentById(long id);
 
     @Transactional
     @Modifying(clearAutomatically = true)
     @Query("update Apartment apartment set apartment.status =:apartmentStatus where apartment.id =:id")
-    void updateApartmentStatusById(@Param("id") int id, @Param("apartmentStatus") ApartmentStatus apartmentStatus);
+    void updateApartmentStatusById(@Param("id") long id, @Param("apartmentStatus") ApartmentStatus apartmentStatus);
 
 
     List<Apartment> findApartmentsByApartmentClassAndPlacesAndStatus(ApartmentClass apartmentClass,
